move state table to stream/common
