+++ src/stream/src/common/mod.rs
@@ -18,12 +18,13 @@ pub use infallible_expr::*;
 use risingwave_common::row::Row;
 use risingwave_storage::StateStore;
 
-use crate::executor::table::state_table::{RowStream, StateTable};
+use self::table::state_table::{RowStream, StateTable};
 use crate::executor::StreamExecutorResult;
 
 mod builder;
 mod column_mapping;
 mod infallible_expr;
+pub mod table;
 
 pub async fn iter_state_table<'a, S: StateStore>(
     state_table: &'a StateTable<S>,
