+++ src/stream/src/common/table/test_batch_table.rs
@@ -23,7 +23,7 @@ use risingwave_storage::memory::MemoryStateStore;
 use risingwave_storage::table::batch_table::storage_table::StorageTable;
 use risingwave_storage::table::{Distribution, TableIter};
 
-use crate::executor::table::state_table::StateTable;
+use crate::common::table::state_table::StateTable;
 
 /// There are three struct in relational layer, StateTable, MemTable and CellBasedTable.
 /// `StateTable` provides read/write interfaces to the upper layer streaming operator.
