+++ src/stream/src/executor/aggregation/minput.rs
@@ -34,8 +34,8 @@ use super::state_cache::extreme::ExtremeAgg;
 use super::state_cache::string_agg::StringAgg;
 use super::state_cache::{CacheKey, GenericStateCache, StateCache};
 use super::AggCall;
+use crate::common::table::state_table::StateTable;
 use crate::common::{iter_state_table, StateTableColumnMapping};
-use crate::executor::table::state_table::StateTable;
 use crate::executor::{PkIndices, StreamExecutorResult};
 
 /// Aggregation state as a materialization of input chunks.
@@ -297,9 +297,9 @@ mod tests {
     use risingwave_storage::StateStore;
 
     use super::MaterializedInputState;
+    use crate::common::table::state_table::StateTable;
     use crate::common::StateTableColumnMapping;
     use crate::executor::aggregation::{AggArgs, AggCall};
-    use crate::executor::table::state_table::StateTable;
     use crate::executor::StreamExecutorResult;
 
     fn create_chunk<S: StateStore>(
