+++ src/stream/src/executor/dynamic_filter.rs
@@ -34,10 +34,10 @@ use super::barrier_align::*;
 use super::error::StreamExecutorError;
 use super::managed_state::dynamic_filter::RangeCache;
 use super::monitor::StreamingMetrics;
-use super::table::state_table::StateTable;
 use super::{
     ActorContextRef, BoxedExecutor, BoxedMessageStream, Executor, Message, PkIndices, PkIndicesRef,
 };
+use crate::common::table::state_table::StateTable;
 use crate::common::{InfallibleExpression, StreamChunkBuilder};
 use crate::executor::expect_first_barrier_from_aligned_stream;
