+++ src/stream/src/executor/lookup/sides.rs
@@ -22,8 +22,8 @@ use risingwave_common::types::DataType;
 use risingwave_common::util::sort_util::OrderPair;
 use risingwave_storage::StateStore;
 
+use crate::common::table::state_table::StateTable;
 use crate::executor::error::StreamExecutorError;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::{Barrier, Executor, Message, MessageStream};
 
 /// Join side of Lookup Executor's stream
