+++ src/stream/src/executor/managed_state/dynamic_filter.rs
@@ -27,8 +27,8 @@ use risingwave_common::types::{ScalarImpl, VIRTUAL_NODE_SIZE};
 use risingwave_common::util::epoch::EpochPair;
 use risingwave_storage::StateStore;
 
+use crate::common::table::state_table::{prefix_range_to_memcomparable, StateTable};
 use crate::executor::error::StreamExecutorError;
-use crate::executor::table::state_table::{prefix_range_to_memcomparable, StateTable};
 use crate::executor::StreamExecutorResult;
 
 type ScalarRange = (Bound<ScalarImpl>, Bound<ScalarImpl>);
