+++ src/stream/src/executor/managed_state/join/mod.rs
@@ -38,9 +38,9 @@ use risingwave_storage::StateStore;
 
 use self::iter_utils::zip_by_order_key;
 use crate::cache::{cache_may_stale, EvictableHashMap, ExecutorCache, LruManagerRef};
+use crate::common::table::state_table::StateTable;
 use crate::executor::error::StreamExecutorResult;
 use crate::executor::monitor::StreamingMetrics;
-use crate::executor::table::state_table::StateTable;
 use crate::task::ActorId;
 
 type DegreeType = u64;
