+++ src/stream/src/executor/managed_state/top_n/top_n_state.rs
@@ -21,8 +21,8 @@ use risingwave_connector::source::DataType;
 use risingwave_storage::StateStore;
 
 use crate::common::iter_state_table;
+use crate::common::table::state_table::StateTable;
 use crate::executor::error::StreamExecutorResult;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::top_n::{serialize_pk_to_cache_key, CacheKey, TopNCache};
 
 /// * For TopN, the storage key is: `[ order_by + remaining columns of pk ]`
