+++ src/stream/src/executor/test_utils.rs
@@ -157,9 +157,9 @@ pub mod agg_executor {
     use risingwave_expr::expr::AggKind;
     use risingwave_storage::StateStore;
 
+    use crate::common::table::state_table::StateTable;
     use crate::common::StateTableColumnMapping;
     use crate::executor::aggregation::{AggCall, AggStateStorage};
-    use crate::executor::table::state_table::StateTable;
     use crate::executor::{
         ActorContextRef, BoxedExecutor, Executor, GlobalSimpleAggExecutor, PkIndices,
     };
@@ -327,7 +327,7 @@ pub mod top_n_executor {
     use risingwave_common::util::sort_util::OrderType;
     use risingwave_storage::memory::MemoryStateStore;
 
-    use crate::executor::table::state_table::StateTable;
+    use crate::common::table::state_table::StateTable;
     pub async fn create_in_memory_state_table(
         data_types: &[DataType],
         order_types: &[OrderType],
