+++ src/stream/src/executor/top_n/top_n_appendonly.rs
@@ -25,10 +25,10 @@ use risingwave_storage::StateStore;
 
 use super::utils::*;
 use super::TopNCache;
+use crate::common::table::state_table::StateTable;
 use crate::error::StreamResult;
 use crate::executor::error::StreamExecutorResult;
 use crate::executor::managed_state::top_n::ManagedTopNState;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::{ActorContextRef, Executor, ExecutorInfo, PkIndices, PkIndicesRef};
 
 /// If the input contains only append, `AppendOnlyTopNExecutor` does not need
