+++ src/stream/src/executor/top_n/top_n_plain.rs
@@ -24,10 +24,10 @@ use risingwave_storage::StateStore;
 
 use super::utils::*;
 use super::{TopNCache, TopNCacheTrait};
+use crate::common::table::state_table::StateTable;
 use crate::error::StreamResult;
 use crate::executor::error::StreamExecutorResult;
 use crate::executor::managed_state::top_n::ManagedTopNState;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::{ActorContextRef, Executor, ExecutorInfo, PkIndices, PkIndicesRef};
 
 /// `TopNExecutor` works with input with modification, it keeps all the data
