+++ src/stream/src/executor/watermark_filter.rs
@@ -29,10 +29,10 @@ use risingwave_storage::StateStore;
 
 use super::error::StreamExecutorError;
 use super::filter::SimpleFilterExecutor;
-use super::table::state_table::StateTable;
 use super::{
     ActorContextRef, BoxedExecutor, Executor, ExecutorInfo, Message, StreamExecutorResult,
 };
+use crate::common::table::state_table::StateTable;
 use crate::common::InfallibleExpression;
 use crate::executor::{expect_first_barrier, Watermark};
