+++ src/stream/src/from_proto/global_simple_agg.rs
@@ -16,8 +16,8 @@
 
 use super::agg_common::{build_agg_call_from_prost, build_agg_state_storages_from_proto};
 use super::*;
+use crate::common::table::state_table::StateTable;
 use crate::executor::aggregation::AggCall;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::GlobalSimpleAggExecutor;
 
 pub struct GlobalSimpleAggExecutorBuilder;
