+++ src/stream/src/from_proto/hash_agg.rs
@@ -22,9 +22,9 @@ use risingwave_common::types::DataType;
 use super::agg_common::{build_agg_call_from_prost, build_agg_state_storages_from_proto};
 use super::*;
 use crate::cache::LruManagerRef;
+use crate::common::table::state_table::StateTable;
 use crate::executor::aggregation::{AggCall, AggStateStorage};
 use crate::executor::monitor::StreamingMetrics;
-use crate::executor::table::state_table::StateTable;
 use crate::executor::{ActorContextRef, HashAggExecutor, PkIndices};
 
 pub struct HashAggExecutorDispatcherArgs<S: StateStore> {
