+++ src/stream/src/from_proto/sort.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use super::*;
-use crate::executor::table::state_table::StateTable;
+use crate::common::table::state_table::StateTable;
 use crate::executor::SortExecutor;
 
 pub struct SortExecutorBuilder;
